Moved IHeldPosition interface to avoid cyclic import in future commits
