+++ src/collective/contact/core/browser/excelexport.py
@@ -18,8 +18,7 @@ except ImportError:
     HAS_EXCELEXPORT = False
 
 from collective.contact.widget.interfaces import IContactChoice, IContactContent
-from collective.contact.core.content.held_position import IHeldPosition
-from collective.contact.core.interfaces import IContactable
+from collective.contact.core.interfaces import IContactable, IHeldPosition
 from collective.contact.core.behaviors import ADDRESS_FIELDS
