+++ src/collective/contact/core/interfaces.py
@@ -1,7 +1,13 @@
 from zope.interface import Interface
 from zope import schema
 
+from plone.namedfile.field import NamedImage
+from plone.supermodel import model
+
 from collective.contact.core import _
+from collective.contact.core.schema import ContactChoice
+from collective.contact.widget.interfaces import IContactContent
+from collective.contact.widget.source import ContactSourceBinder
 
 
 class IContactable(Interface):
@@ -62,4 +68,45 @@ class IPersonHeldPositions(Interface):
 
     def get_sorted_positions(self):
         """Get sorted positions
+        """
+
+
+class IHeldPosition(model.Schema, IContactContent):
+    """Interface for HeldPosition content type"""
+
+    position = ContactChoice(
+        title=_("Organization/Position"),
+        source=ContactSourceBinder(portal_type=("organization", "position")),
+        required=True,
+    )
+    label = schema.TextLine(
+        title=_("Additional label"),
+        description=_("Additional label with information that does not appear "
+                      "on position label"),
+        required=False)
+    start_date = schema.Date(
+        title=_("Start date"),
+        required=False,
+    )
+    end_date = schema.Date(
+        title=_("End date"),
+        required=False,
+    )
+    photo = NamedImage(
+        title=_("Photo"),
+        required=False,
+        readonly=True,
+    )
+
+    def get_person(self):
+        """Returns the person who holds the position
+        """
+
+    def get_position(self):
+        """Returns the position (if position field is a position)
+        """
+
+    def get_organization(self):
+        """Returns the first organization related to HeldPosition
+        i.e. position field or parent of the position
         """
